core/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
debugpy==1.6.3
defusedxml==0.7.1
Django==3.2.18
django-cors-headers==3.13.0
django-csp==3.6
django-db-schema-renderer==0.1.2
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.17.0
django-permissions-policy==4.14.0
django-spaghetti-and-meatballs==0.4.2
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-xml==2.0.0
drf-extensions==0.7.1
drf-writable-nested==0.7.0
drf-spectacular==0.26.0
drf_extra_fields==3.4.1
drf_base64
filetype
graphviz==0.20.1
idna==3.4
model-bakery==1.3.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pygraphviz==1.10
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
unicodecsv==0.14.1
urllib3==1.26.14
uWSGI==2.0.21
djangorestframework-simplejwt==4.8.0
python-box[all]~=7.0
